Adjustments to welcome/login screens

+++ PropsApp/src/screens/SignUp.js
@@ -1,10 +1,10 @@
-import { View, Text, ActivityIndicator, Button } from 'react-native'
+import { View, Text, ActivityIndicator, Button, Alert, StyleSheet, TouchableOpacity, Pressable } from 'react-native'
 import React from 'react'
-import { SafeAreaView } from 'react-native-safe-area-context';
+import { SafeAreaView, withSafeAreaInsets } from 'react-native-safe-area-context';
 import { useEffect, useState } from 'react';
 import { getAuth, createUserWithEmailAndPassword } from 'firebase/auth';
 import { TextInput } from 'react-native-gesture-handler';
-import { doc, setDoc} from "firebase/firestore";
+import { doc, setDoc, runTransaction} from "firebase/firestore";
 import { FIRESTORE_DB } from '../config/firebaseConfig';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 const auth = getAuth();
@@ -12,20 +12,27 @@ const auth = getAuth();
 const SignUp = ({ navigation }) => {
     const [email, setEmail] = useState("");
     const [password, setPassword] = useState("");
+    const [confirmPassword, setConfirmPassword] = useState("");
+    const [username, setUsername] = useState("");
     const [isLoading, setIsLoading] = useState(false);
 
     const signUp = async () => {
+        if (password !== confirmPassword){
+          Alert.alert("Error", "Passwords do not match.");
+          return;
+        }
         setIsLoading(true);
         try{
             const userCred = await createUserWithEmailAndPassword(auth, email, password);
             const uid = userCred.user.uid;
+            await addUsernameToUser(uid);
             await AsyncStorage.setItem('UserUID', uid);
             //Add user data to the database
             //Will overwrite existing data if a user is deleted
             await setDoc(doc(FIRESTORE_DB, 'users', uid), {
                 email: email,
                 uid: uid,
-            });
+            }, {merge:true});
         }catch(err){
             console.log(err);
             alert("Unknown Error: " + err);
@@ -34,6 +41,26 @@ const SignUp = ({ navigation }) => {
         }
     }
 
+    const addUsernameToUser = async (uid) => {
+
+      const usernamesRef = doc(FIRESTORE_DB, 'usernames', username);
+      const userRef = doc(FIRESTORE_DB, 'users', uid);
+
+      try {
+          await runTransaction(FIRESTORE_DB, async (transaction) => {
+              const usernameDoc = await transaction.get(usernamesRef);
+              if (usernameDoc.exists()){
+                  throw new Error("Username already taken.");
+              }
+              transaction.set(usernamesRef, {uid});
+              transaction.set(userRef, {username});
+          });
+          console.log("Username added successfully.");
+      }catch(error){
+          console.error("Error adding username:", error);
+          alert(error.message);
+      } 
+  }
     if (isLoading){
         return (
             <SafeAreaView>
@@ -42,28 +69,90 @@ const SignUp = ({ navigation }) => {
         )
     }
   return (
-    <SafeAreaView>
-      <Text>Sign Up</Text>
+    <SafeAreaView style={styles.container}>
+      <Text style={[styles.text, styles.header]}>Sign Up</Text>
       <TextInput
+        style={styles.textBox}
+        onChangeText={text => setUsername(text)}
+        value={username}
+        placeholder="Username"
+        autoCapitalize="none"
+      ></TextInput>
+      <TextInput
+        style={styles.textBox}
         onChangeText={text => setEmail(text)}
         value={email}
         placeholder="Email"
         autoCapitalize="none"
       ></TextInput>
       <TextInput
+        style={styles.textBox}
         onChangeText={text => setPassword(text)}
         value={password}
         placeholder="Password"
         secureTextEntry={true}
         autoCapitalize="none"
       ></TextInput>
-      <Button
-      title='Sign Up'
+      <TextInput
+        style={styles.textBox}
+        onChangeText={text => setConfirmPassword(text)}
+        value={confirmPassword}
+        placeholder="Confirm Password"
+        secureTextEntry={true}
+        autoCapitalize="none"
+      ></TextInput>
+      <TouchableOpacity
+      style={[styles.button]}
       onPress={signUp}
-      />
+      >
+      <Text style={styles.text}>Sign Up</Text>  
+      </TouchableOpacity>
+      
+      <View>
+        <Pressable onPress={() => navigation.navigate('Sign In')}>
+          <Text style={styles.text}>Already have an account? Sign in <Text style={styles.blue}>here.</Text></Text>
+        </Pressable>
+      </View>
 
     </SafeAreaView>
   )
 }
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: '#1a1a1a'
+  },
+  text: {
+    color: '#e8e8e8',
+    fontSize: 18
+  },
+  header:{
+    fontWeight: 'bold',
+    marginTop: 20
+  },
+  textBox:{
+    width: '80%',
+    padding: 20,
+    margin: 10,
+    borderWidth: 2,
+    borderRadius: 10,
+    backgroundColor: 'white'
+  },
+  button:{
+    width: '80%',
+    padding: 20,
+    margin: 10,
+    borderWidth: 2,
+    borderRadius: 10,
+    backgroundColor: '#4a92ff',
+    alignItems: 'center',
+  },
+  blue:{
+    color: '#4a92ff',
+  }
+});
 
 export default SignUp
